ci: run build.sh in host environment
Previously we tried to install docker inside a debian:sid container
while the host is ubuntu:trusty. This brings an obvious issue that the
docker version may not be the same in between, and docker client/server
must match to communicate to each other.

This changes allows us to copy debian qemu-user-static binaries in the
container, while run the docker build process in the host side.